+++ generators/common/inputs/isThisYou.js
@@ -23,22 +23,32 @@
 
 	module.exports = class AppName {
 		static set($that) {
-			logs.hintHeader($that);
+			if ($that.config.get('authorEmail')) {
+				logs.hintHeader($that);
 
-			prompts[0].message += $that.config.get('authorFirstName');
-			prompts[0].message += ' ';
-			prompts[0].message += $that.config.get('authorLastName');
-			prompts[0].message += ' <';
-			prompts[0].message += $that.config.get('authorEmail');
-			prompts[0].message += '> ?';
+				prompts[0].message += $that.config.get('authorFirstName');
+				prompts[0].message += ' ';
+				prompts[0].message += $that.config.get('authorLastName');
+				prompts[0].message += ' <';
+				prompts[0].message += $that.config.get('authorEmail');
+				prompts[0].message += '> ?';
 
-			return $that.prompt(prompts).then($response => {
-				this.onSuccess($that, $response);
-			});
+				return $that.prompt(prompts).then($response => {
+					this.onSuccess($that, $response);
+				});
+			}
+			this.onError($that);
+			return null;
 		}
 
+		/* istanbul ignore next */
 		static onSuccess($that, $response) {
 			$that.isThisYou = $response.isThisYou;
+			$that.log();
+		}
+
+		static onError($that) {
+			$that.isThisYou = false;
 			$that.log();
 		}
 	};