+++ generators/config/index.js
@@ -1,0 +1,76 @@
+/**
+ * Generated header by C0ZEN for generator-cozen-angular project
+ * Generated file index on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 20/12/2017
+ * Time: 20:49
+ * Version: 1.0.0
+ */
+(function () {
+	'use strict';
+
+	const Generator     = require('yeoman-generator');
+	const chalkInstance = require('chalk');
+	const chalk         = new chalkInstance.constructor({
+		level  : 3,
+		enabled: true
+	});
+
+	const colors = require('../common/colors.js');
+	const logs   = require('../common/logs.js');
+
+	const appName = require('./inputs/appName.js');
+
+	module.exports = class extends Generator {
+		initializing() {
+			logs.priority(this, 'initializing', true);
+
+			const prompts = [
+				{
+					type   : 'confirm',
+					name   : 'areYouReady',
+					message: 'Are you ready to set or update your current generator config ?',
+					default: true
+				}
+			];
+
+			return this.prompt(prompts).then($response => {
+				this.areYouReady = $response.areYouReady;
+				if (!this.areYouReady) {
+					this.env.error('You are not ready !');
+				}
+			});
+		}
+
+		prompting() {
+			logs.priority(this, 'prompting', true);
+		}
+
+		configuring() {
+			logs.priority(this, 'configuring', false);
+		}
+
+		inputAppName() {
+			return appName.set(this);
+		}
+
+		writing() {
+			logs.priority(this, 'writing', false);
+		}
+
+		conflicts() {
+			logs.priority(this, 'conflicts', false);
+		}
+
+		install() {
+			logs.priority(this, 'install', false);
+		}
+
+		end() {
+			logs.priority(this, 'end', false);
+			this.log(chalk.hex(colors.get('green'))('Your config is ready ;)'));
+		}
+	};
+
+})();