+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+## [0.7.0]
+### Added 
+- Space before INITIALIZING step
+- Config in the package to globalize the common data (appName) to avoid requesting it when the `.yo-rc.json` file is not present locally (which is often due to gitignore - in a good way -). So know we can assume that there is a global config and a local config.
+- Merge between global and local config (the global config override the local config only when the local config properties does not exist)
+- New sub-generator `filter` to create an angular
+- New sub-generator `config` to update the config
+
+### Changed
+- When the user is not set in the config, the isThisYou prompt will be automatically skipped
+- The areYouReady prompt show the real app name instead of the camel one
+- Updated the readme about new sub-generators
 
 ## [0.6.0]
 ### Changed